+++ src/Symfony/Component/DependencyInjection/Loader/DirectoryLoader.php
@@ -11,8 +11,6 @@
 
 namespace Symfony\Component\DependencyInjection\Loader;
 
-use Symfony\Component\Config\Loader\LoaderInterface;
-
 /**
  * DirectoryLoader is a recursive loader to go through directories.
  *
@@ -37,7 +35,7 @@ class DirectoryLoader extends FileLoader
 
                 $this->setCurrentDir($path);
 
-                $this->import($dir, null, LoaderInterface::ERROR_LEVEL_ALL, $path);
+                $this->import($dir, null, parent::IGNORE_NONE, $path);
             }
         }
     }