+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -3,8 +3,9 @@ CHANGELOG
 
 4.4.0
 -----
-* added option `ignore_not_found` for imported config files
-* deprecated 3rd `boolean` argument `$ignoreErrors` and changed to `int` `$errorLevel` in `import` method  in `FileLoader`.
+
+ * added option `ignore_not_found` for imported config files
+ * deprecated passing a boolean to argument `$ignoreErrors` of `FileLoader::import()`, pass an `int` instead
 
 4.3.0
 -----